requirements.txt
bidict==0.22.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-SocketIO==5.3.6
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyserial==3.5
python-engineio==4.8.0
python-socketio==5.10.0
rplidar-roboticia==0.9.5
simple-websocket==1.0.0
Werkzeug==3.0.1
wsproto==1.2.0